+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
+#### 0.1.4.wip - MAU `2016-12-12`  
+**Added methods to I18n class**
+
+* Added All I18n methods from old gem, fully documented and tested
+* Cleaned up some other documentation
+
+---
+
 #### 0.1.3.wip - MAU `2016-12-12`  
 **Added methods to Hash class**
 
 * Added All Hash methods from old gem, fully documented and tested
 
 ---
+
 #### 0.1.2.wip - MAU `2016-12-10`  
 **Added methods to File and add a new module to support testing needs**